day127/database_audit_collector/requirements.txt
# Database connectors
psycopg2-binary==2.9.9
pymongo==4.7.2
mysql-connector-python==8.4.0
pyodbc==5.1.0

# Async and concurrency
asyncio==3.4.3
asyncpg==0.29.0
aioredis==2.0.1

# Web framework and API
fastapi==0.111.0
uvicorn==0.30.1
websockets==12.0
jinja2==3.1.4

# Data processing
pydantic==2.7.1
jsonschema==4.22.0
python-dateutil==2.9.0

# Logging and monitoring
structlog==24.1.0
prometheus-client==0.20.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0

# Security
cryptography==42.0.7
bcrypt==4.1.3

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1
